[refactor] bookmark.d.ts 파일 생성,  인터페이스와 타입코드 이동
