+++ src/components/bookmark/BookmarkedEachItem.tsx
@@ -2,14 +2,6 @@ import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { bookmarkDataInterface } from '../../hooks/useBookmarks';
 
-// interface GridItemProps {
-//   id: string;
-//   url : string;
-//   img : string;
-//   title : string;
-// }
-
-
 const BookmarkedEachItem = ({bookmarkId,bookmarkOgImg,bookmarkTitle} : bookmarkDataInterface) => {
   
   const { attributes, listeners, setNodeRef, transform, transition } = useSortable({id : bookmarkId });
