+++ src/types/bookmark.d.ts
@@ -1,0 +1,16 @@
+
+export interface manageModalInterface {
+    closeModal : () => void;
+    handleCheckedItems : (items : bookmarkDataInterface[]) => void;
+    gridItems?: bookmarkDataInterface[]; // 부모로부터 받은 그리드 항목
+}
+
+
+
+export interface bookmarkDataInterface {
+    bookmarkId : string;
+    bookmarkTitle : string;
+    bookmarkURL : string;
+    bookmarkOgImg : string;
+    checked ? : boolean;
+}
